update client tests and add api tests
